% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canyonb-package.R
\docType{package}
\name{canyonb-package}
\alias{canyonb}
\alias{canyonb-package}
\title{canyonb: Estimate Open Ocean CO2 Variables and Macronutrients from T, S and O2 data}
\description{
An alternative to static climatologies: Robust estimation of open ocean CO2 variables and nutrient concentrations from T, S and O2 data using Bayesian neural networks.
}
\author{
\strong{Maintainer}: Henry Bittig \email{henry.bittig@io-warnemuende.de}

}
\keyword{internal}
