% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgts_AT}
\alias{wgts_AT}
\title{AT variable}
\format{
A data frame with AT variables:
}
\usage{
wgts_AT
}
\description{
A dataset containing the AT variable
The variables are as follows:
}
\keyword{datasets}
