% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgts_NO3}
\alias{wgts_NO3}
\title{NO3 variable}
\format{
A data frame with NO3 variables:
}
\usage{
wgts_NO3
}
\description{
A dataset containing the NO3 variable
The variables are as follows:
}
\keyword{datasets}
