% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgts_pC02}
\alias{wgts_pC02}
\title{pCO2 variable}
\format{
A data frame with pCO2 variables:
}
\usage{
wgts_pC02
}
\description{
A dataset containing the pCO2 variable
The variables are as follows:
}
\keyword{datasets}
