% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgts_PO4.R
\docType{data}
\name{wgts_PO4}
\alias{wgts_PO4}
\title{(Data) wgts_PO4.rda}
\format{
An object of class \code{data.frame} with 781 rows and 29 columns.
}
\usage{
wgts_PO4
}
\description{
Some description on the data
}
\details{
abstract from research article:

some abstract ??
}
\examples{
data(wgts_PO4)
}
\references{
Matlab and R routines to robustly estimate open ocean CO2 variables and macronutrients from T, S and O2 data using Bayesian neural networks ("CANYON-B") of Bittig et al. (2018) Front. Mar. Sci. 5:328. http://dx.doi.org/10.3389/fmars.2018.00328
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Henry Bittig
}\if{html}{\out{</div>}}
}
\keyword{data}
