% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgts_CT}
\alias{wgts_CT}
\title{CT variable}
\format{
A data frame with CT variables:
}
\usage{
wgts_CT
}
\description{
A dataset containing the CT variable
The variables are as follows:
}
\keyword{datasets}
