% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgts_pH}
\alias{wgts_pH}
\title{pH variable}
\format{
A data frame with pH variables:
}
\usage{
wgts_pH
}
\description{
A dataset containing the pH variable
The variables are as follows:
}
\keyword{datasets}
