% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgts_PO4}
\alias{wgts_PO4}
\title{PO4 variable}
\format{
A data frame with PO4 variables:
}
\usage{
wgts_PO4
}
\description{
A dataset containing the PO4 variable
The variables are as follows:
}
\keyword{datasets}
