% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgts_SiOH4}
\alias{wgts_SiOH4}
\title{SiOH4 variable}
\format{
A data frame with SiOH4 variables:
}
\usage{
wgts_SiOH4
}
\description{
A dataset containing the SiOH4 variable
The variables are as follows:
}
\keyword{datasets}
